% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.r
\name{loadings}
\alias{loadings}
\alias{print.loadings}
\alias{print.fad}
\title{Print Loadings in Factor Analysis using fad.}
\usage{
loadings(x, ...)

\method{print}{loadings}(x, digits = 3L, cutoff = 0.1, sort = FALSE,
  ...)

\method{print}{fad}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class "fad" or the loadings component of a "fad" object.}

\item{\dots}{further arguments for other methods, ignored for loadings.}

\item{digits}{number of decimal places to use in printing uniquenesses and loadings.}

\item{cutoff}{loadings smaller than this (in absolute value) are suppressed.
Each variable with any loading larger than 0.5 (in modulus) is assigned to
the factor with the largest loading, and the variables are printed in the order of
the factor they are assigned to, then those unassigned.}

\item{sort}{logical. If true, the variables are sorted by their importance on each factor.}

\item{loadings}{smaller than this (in absolute value) are suppressed.}
}
\description{
Extract or print loadings in factor analysis from package fad.
}
