% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.fads}
\alias{print.fads}
\title{Print the Output of Factor Analysis}
\usage{
\method{print}{fads}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class \code{fads}.}

\item{digits}{number of decimal places to use in printing uniquenesses and loadings.}

\item{\dots}{further arguments to \code{print}.}
}
\value{
None.
}
\description{
Prints the output of the \code{fads}.
}
